python_django/requirements.txt
Django == 3.1.2
djangorestframework == 3.12.1
PyHamcrest == 2.0.2